+++ general/fight/FightLoader.h
@@ -13,13 +13,13 @@ class FightLoader
 {
 public:
     static QList<Type> loadTypes(const QString &file);
-    static QHash<quint32,Monster> loadMonster(const QString &folder, const QHash<quint32,Skill> &monsterSkills, const QList<Type> &types, const QHash<quint32, Item> &items);
-    static QHash<quint32/*item*/, QHash<quint32/*monster*/,quint32/*evolveTo*/> > loadMonsterEvolutionItems(const QHash<quint32,Monster> &monsters);
-    static QHash<quint32/*item*/, QSet<quint32/*monster*/> > loadMonsterItemToLearn(const QHash<quint32,Monster> &monsters, const QHash<quint32/*item*/, QHash<quint32/*monster*/,quint32/*evolveTo*/> > &evolutionItem);
-    static QHash<quint32,Skill> loadMonsterSkill(const QString &folder, const QHash<quint32,Buff> &monsterBuffs, const QList<Type> &types);
-    static QHash<quint32,Buff> loadMonsterBuff(const QString &folder);
-    static QHash<quint32,BotFight> loadFight(const QString &folder, const QHash<quint32,Monster> &monsters, const QHash<quint32, Skill> &monsterSkills, const QHash<quint32, Item> &items);
-    static QList<PlayerMonster::PlayerSkill> loadDefaultAttack(const quint32 &monsterId, const quint8 &level, const QHash<quint32,Monster> &monsters, const QHash<quint32, Skill> &monsterSkills);
+    static QHash<quint16,Monster> loadMonster(const QString &folder, const QHash<quint16,Skill> &monsterSkills, const QList<Type> &types, const QHash<quint16, Item> &items);
+    static QHash<quint16/*item*/, QHash<quint16/*monster*/,quint16/*evolveTo*/> > loadMonsterEvolutionItems(const QHash<quint16,Monster> &monsters);
+    static QHash<quint16/*item*/, QSet<quint16/*monster*/> > loadMonsterItemToLearn(const QHash<quint16,Monster> &monsters, const QHash<quint16/*item*/, QHash<quint16/*monster*/,quint16/*evolveTo*/> > &evolutionItem);
+    static QHash<quint16,Skill> loadMonsterSkill(const QString &folder, const QHash<quint8,Buff> &monsterBuffs, const QList<Type> &types);
+    static QHash<quint8,Buff> loadMonsterBuff(const QString &folder);
+    static QHash<quint16,BotFight> loadFight(const QString &folder, const QHash<quint16,Monster> &monsters, const QHash<quint16, Skill> &monsterSkills, const QHash<quint16, Item> &items);
+    static QList<PlayerMonster::PlayerSkill> loadDefaultAttack(const quint16 &monsterId, const quint8 &level, const QHash<quint16,Monster> &monsters, const QHash<quint16, Skill> &monsterSkills);
     static QString text_type;
     static QString text_name;
     static QString text_id;
